@@ -1,5 +1,7 @@
 module lotto
 
-go 1.12
+go 1.24.0
 
 require github.com/sirupsen/logrus v1.9.3
+
+require golang.org/x/sys v0.36.0 // indirect
