@@ -10,6 +10,15 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+// random is a random number generator
+var random *rand.Rand
+
+// init initializes the random number generator
+func init() {
+	random = rand.New(rand.NewSource(time.Now().UnixNano()))
+}
+
+// createSlice creates a slice of integers from 1 to max
 func createSlice(max int) []int {
 	min := 1
 	arr := make([]int, max)
@@ -20,19 +29,33 @@ func createSlice(max int) []int {
 	return arr
 }
 
+// generateNumbers generates n unique random numbers from 1 to l
 func generateNumbers(n int, l int) []int {
 	numbrs := createSlice(l)
-	rand.Seed(time.Now().UnixNano())
 	rand.Shuffle(len(numbrs), func(i, j int) { numbrs[i], numbrs[j] = numbrs[j], numbrs[i] })
 
 	return numbrs[0:n]
 }
 
+// sortNumbers sorts a slice of integers in ascending order
+func sortNumbers(numbers []int) []int {
+	// Bubble sort
+	for i := range len(numbers) {
+		for j := 0; j < len(numbers)-i-1; j++ {
+			if numbers[j] > numbers[j+1] {
+				numbers[j], numbers[j+1] = numbers[j+1], numbers[j]
+			}
+		}
+	}
+	return numbers
+}
+
+// generateNumbersString generates a string of n unique random numbers from 1 to l
 func generateNumbersString(n int, l int) string {
-	numbrs := generateNumbers(n, l)
+	numbrs := sortNumbers(generateNumbers(n, l))
 	var txt string
 
-	for i := 0; i < n; i++ {
+	for i := range len(numbrs) {
 		if i == 0 {
 			txt = strconv.Itoa(numbrs[i])
 		} else {
@@ -42,10 +65,13 @@ func generateNumbersString(n int, l int) string {
 	return txt
 }
 
+// the main function
 func main() {
 	arg, _ := strconv.Atoi(os.Args[1])
+
 	log.SetFormatter(&log.JSONFormatter{})
-	rand.Seed(time.Now().UnixNano())
+	//log.SetLevel(log.DebugLevel)
+	//log.Debugf("Argument: %v", arg)
 
 	fmt.Printf("Set of %v.\n", arg)
 	for i := 1; i < arg+1; i++ {
